Move android port code to servoshell
