+++ etc/run_in_headless_android_emulator.py
@@ -66,7 +66,7 @@ def main(avd_name, apk_path, *args):
         # in case they say something useful while we wait in subsequent steps.
         logcat_args = [
             "--format=raw",  # Print no metadata, only log messages
-            "simpleservo:D",  # Show (debug level) Rust stdio
+            "servoshell:D",  # Show (debug level) Rust stdio
             "*:S",  # Hide everything else
         ]
         with terminate_on_exit(adb + ["logcat"] + logcat_args) as logcat:
