+++ ports/servoshell/egl/mod.rs
@@ -1,0 +1,9 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#[cfg(target_os = "android")]
+pub mod gl_glue;
+
+#[cfg(target_os = "android")]
+mod android;
